This initial commit includes a working copy of the Coffee Pot code.
This code controls a valve to allow water to refill the tank in
a Keurig coffee pot based on input from a float valve.

If the valve is turned on or off, it is held that way for at least
2 seconds as a debounce mechanism. If the valve is kept on for 20
seconds, turn it off until we are reset. This is to prevent
errors from flooding the kitchen.

There is an LED that is turned on when the valve is open, and flashed
when we are in 'safety off'.